requirements.txt
aiogram==3.4.1
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.0
python-dotenv==1.0.0
openpyxl==3.1.2
aiohttp>=3.9.0
passlib[bcrypt]==1.7.4
gunicorn==21.2.0
